+++ fusion/src/Fusion/autodiff/Sort.h
@@ -8,20 +8,20 @@
 template <typename T> class Sort {
 public:
   std::vector<NodeID> sorted;
-  Sort(uint16_t numNodes) : numNodes_(std::move(numNodes)) {
+  Sort(size_t numNodes) : numNodes_(std::move(numNodes)) {
     std::vector<bool> visited(numNodes, false);
   };
 
-  std::vector<uint16_t> calc_indegree(std::vector<INode<T>> &nodes,
+  std::vector<size_t> calc_indegree(std::vector<INode<T>> &nodes,
                                       std::vector<ProducerInfo> &produced_by) {
     // This belongs in the graph??? not in the sort - also iteratively update
     // this do not calculate on method exe
-    std::vector<uint16_t> inDegree(numNodes_);
-    for (uint16_t i = 0; i < numNodes_; i++) {
-      uint16_t increment = 0;
+    std::vector<size_t> inDegree(numNodes_);
+    for (size_t i = 0; i < numNodes_; i++) {
+      size_t increment = 0;
       auto inp = nodes[i].inputs;
       // inspect O(?) for this impl
-      for (uint16_t j = 0; j < inp.size(); j++) {
+      for (size_t j = 0; j < inp.size(); j++) {
         if (produced_by[inp[j].idx].nid.idx != kNoNode) {
           increment++;
         }
@@ -37,8 +37,8 @@ public:
                    std::vector<std::vector<ConsumerInfo>> &consumed_by,
                    std::vector<NodeID> &node_ids) {
     std::queue<NodeID> q;
-    std::vector<uint16_t> in_degree = calc_indegree(nodes, produced_by);
-    for (uint16_t i = 0; i < numNodes_; i++) {
+    std::vector<size_t> in_degree = calc_indegree(nodes, produced_by);
+    for (size_t i = 0; i < numNodes_; i++) {
       if (i > in_degree.size()) {
         throw std::runtime_error("Number of nodes greater than indegree vec");
       };
@@ -52,9 +52,9 @@ public:
       q.pop();
       result.push_back(nid);
       auto outputs = nodes[nid.idx].outputs;
-      for (uint16_t j = 0; j < outputs.size(); ++j) {
+      for (size_t j = 0; j < outputs.size(); ++j) {
         std::vector<ConsumerInfo> children = consumed_by[outputs[j].idx];
-        for (uint16_t k = 0; k < children.size(); ++k) {
+        for (size_t k = 0; k < children.size(); ++k) {
           in_degree[children[k].nid.idx]--;
           if (in_degree[children[k].nid.idx] == 0) {
             q.push(children[k].nid);
@@ -71,7 +71,7 @@ public:
   }
 
 private:
-  uint16_t numNodes_;
+  size_t numNodes_;
 };
 
 #endif // SORT_H
